Update Zendesk client for new form
We are switching to use a new Zendesk form, so this changes the Zendesk
client. We now need to pass in the ID of the new form, and instead of
populating the "Ticket category" dropdown we want to populate a dropdown
called "Task type". This will only ever have one thing selected, so
doesn't need to be able to take a list of values.