use phonenumbers to validate uk landlines
we already use phonenumbers in format_phone_number_human_readable as it
nicely formats numbers - but hang on, it has a whole suite of validation
features too.

We often find other teams asking us how we validate phone numbers so
they can replicate it on their own frontend. We link to this repo but it
would be much nicer if we could just point them to a well supported
third party library with a variety of implementations in different
languages.

Start by using phonenumbers in a separate function so we can compare the
test output.

We could potentially run validation in parallel with our regular output
to ensure it matches. I'm not too worried if things we currently fail
are allowed in by phonenumbers, as the chances are they're more correct
than we are - but it's important that any numbers we currently accept
and deliver to also work in phonenumbers.