add code to handle international numbers without + or 00 prefix
We have seen excel interpret cells with phone numbers in as numeric
fields (rather than strings). It might strip leading zeros or + prefixes
as they are implied with actual countable numbers. But they're a crucial
part of identifiying international phone numbers.

the phonenumbers library doesn't check for this. If you pass in a number
without an international prefix, it assumes you're trying to call a
local (GB) number, so validates it based on that.

We're more likely to be sending messages to UK numbers, so start with
that, but if the number isn't a valid UK number, AND the service can
send to international numbers, then we can try and parse the number
again but this time with a + on the beginning to force it to identify as
a foreign number.

If this parsing still doesn't work, then the number is clearly invalid.
But if the number works, then use that with the + as the root number
that we can send to later.