change UKLandline to PhoneNumber and make it instanced
previously, UKLandline was a "namespace" style class that just contained
a set of static methods

instead, make it an instanced class where the constructor parses and
validates the phone number.

This'll make the interface cleaner for if we want to, eg, return the
number in different formats, or extend the class to handle different
cases (eg landline vs not, international vs not, strict validation vs
not) later.