Update test cases to use semantically valid phone numbers
Lots of the UK phone number tests used "07123456789". This isn't
actually a valid number, as the "0712" first part of the phone number
hasn't been assigned to any carriers by ofcom[^1].

Update these to one that works just to get the tests to pass.

Note this also means that numbers reserved for TV usage (07700900xxx)
are no longer valid! This may have an impact on our users and perhaps on
our functional tests if we expect these to be delivered - as with the
new code logic they would now fail validation as they're accurately not
valid phone numbers.

[^1]: https://github.com/google/libphonenumber/blob/master/resources/carrier/en/44.txt