Add `unsubscribe_link` argument to email templates
We have started sending `list-unsubscribe` headers if a user marks a
template as unsubscribeable, for teams who aren’t using the API.

However we are already telling these teams to add unsubscribe links to
the body of their emails.

This means:
- there are two different ways to add unsubscribe links
- user will have to manage requests to unsubscribe coming from two
  different places

This will be confusing, and make it harder for us to write good guidance
that users will be able to follow, and harder to explain what the
‘let users unsubscribe from these emails’ checkbox does.

If we have a confusing product proposition and users cant follow our
guidance then we risk our spam score rising.

This commit adds a way to insert an unsubscribe link (which we will
generate) to the footer of every email. The API can then start
generating these links, and the admin app can start responding to them.

We can then report unsubscribes back to the team in the same way, no
matter if the recipient has clicked the button in their email client or
the link at the end of the email.