Added allow_landline argument to PhoneNumber constructor
The default behaviour of notifications_utils.recipient_validation.phone_number.PhoneNumber validates landline numbers. We want to be able to toggle this on or off via a constructor argument so that this code can be used to validate phone numbers for all services, not just those with the sms_to_uk_landline permission. This change adds the constructor argument allow_landline and makes changes to the validation code to throw an exception if a valid landline is passed when allow_landline=False.