Separate out parsing of possible numbers and validation of allowed numbers
Previously, the init method on PhoneNumbers objects did both parsing and validation.
This causes a number of small problems. Firstly, it makes api code difficult to read. There are instances where instansiation of a PhoneNumber object is being used to validate a number against a services permissions; this isn't clear from the code. This also makes the class much less flexible because it becomes tightly bound to a given services permissions.