Add a new custom field to the Zendesk client
The Zendesk client now takes an optional argument called
`user_created_at`. If provided, this value is used to populate a new
field we have on the form which is a calendar date picker. This will let
us analyse how long the people opening tickets have been using Notify for.