version_tools: use importlib.resources for copying config files
move actual copies of these files into the version_tools
sub-package so they can be included in MANIFEST.in

requirements_for_test.txt here becomes requirements_for_test.in
and it is suggested that apps using utils' requirements_for_test
should be using it to freeze their test dependencies, allowing
pip-sync to be used in their freezing workflow.

i still think this is a weird way to manage dependencies etc.