add EventletTimeoutMiddleware
including eventlet detection that will omit full import if
not using eventlet.

this is needed because gunicorn doesn't do this out-of-the-box
for the eventlet worker. instead the `timeout` configuration
variable controls when the whole worker process will be killed
if it hasn't produced *any* output for a certain time - it
doesn't work on a per-request level.