Replaces invalid numbers for the countries under test with truly valid numbers.
The phonenumbers library is much more accurate with validation, owing to its extensive metadata catalogue. As a result, numbers that passed validation (but were not in fact valid) for the old code now fail. This keeps the country under test identical but ensures the numbers are in fact valid for those countries according to the library's metadata.