Freeze all test requirements
We’re going to roll this out across all the apps, so anyone running the
tests in an app will have all the same dependencies as someone else, not
just the same app dependencies.

Pulling this out of
https://github.com/alphagov/notifications-utils/pull/1155/files
so it’s easier to review separately.

We have to remove `-e .` from `requirements.in` and instead of,
explicitly list `setup.py` in the pip-compile command. Otherwise this
line will cause notifications_utils itself to be included in the
compiled requirements, which is not something we want.