use gh api rather than raw.githubusercontent cdn
We've seen issues where raw.githubusercontent.com has returned 403s,
possibly due to some kind of WAF on github's web cache that is blocking
our requests.

We can instead use github's API to get the same data in a more expected
way. Note that it doesn't need auth for public repos, but you get more
lenient rate limits with auth, so if we still run into issues we might
want to look into that as an option

[^1]: https://docs.github.com/en/rest/repos/contents?apiVersion=2022-11-28#get-repository-content