Revert "Automatically copy config after version bump"
This reverts commit 484d4530885448ef4a61ce811bcefa015bc3e941.

We now do this as part of the freeze-requirements step in the repos.

Also because the script now pulls the config files from the local
install (not over the network) the files it will be pulling will be out
of date at this point (before installing the new utils version).