Allow leading empty columns in CSV files
We were stripping leading commas from the CSV data.

This mean that a file which looks like

```CSV
,,,phone number
,,,07900900123
```

Got turned into
```CSV
phone number
,,,07900900123
```

This meant that the `phone number` header and the phone number itself
were no longer in the same column, causing an unexpected error.

We still want to strip commas from the end of the file, so we remove any
completely empty rows.