fix: unsupported countries were not being rejected
if you parse the phone number "24741111", the parser doesn't know if
it's international, so starts off assuming it has a country code of +44.
Then, as that number isn't a valid UK number, we'd go down the
"forced_international_number" route - and stick a plus on the beginning
and parse again. This would prompt phonenumbers to realise it's not a UK
phone number but actually one from the Ascension Island - and in fact is
a valid number there.

The problem is that we don't support Ascension Island. However, the
unsupported country code check takes place before the forced
international code branch - and compares against the assumed "+44" code.

This led to a bug where the phone number appeared valid, but if you
tried to call `get_international_phone_info` it would KeyError when
trying to find billing multiplier info.

Moving this check to validate ensures that we're checking our business
logic against the final number that we'll try and send to, rather than
an earlier version of the number we may end up manipulating further