Don’t symlink pyproject.toml
When someone tries to access this file from the Github API it doesn’t
cleanly redirect.

This is causing an error with what uv is trying to do in
https://github.com/astral-sh/uv/pull/10765/files

We can do some work later to make sure these 2 files don’t get out of
sync, or some other longer-term fix.