Freeze requirements using `uv`
Same as we do in other repos.

`make freeze-requirements` is only ever run locally so it can depend on
developers having uv installed locally.

Can’t change the equivalent commands in `make bootstrap` yet because in
CI this repo doesn’t run the tests in Docker, so there’s no Dockerfile
in which to install uv.