Replace `setup.py` with `pyproject.toml`
As of https://peps.python.org/pep-0621/ the Python community has
standardised on using `pyproject.toml` for defining a project’s
metadata.

In Python 3.12 setuptools (which we use in `setup.py`) won’t be
installed by default, so we may want to move towards not using it.

For both reasons this change will future-proof our code.

This commit follows the guidance in https://packaging.python.org/en/latest/guides/modernize-setup-py-project/#how-to-handle-packaging-metadata
(including the part about moving all static metadata to
`pyproject.toml`).

Therefore it still uses setuptools for reading the version. This can be
changed later if we want to drop setuptools entirely.