Rename function argument for consistency
Instead of having two similar boolean args, let's just have one.

Since then we decide more based on this arg than whether Notify tag
is included, let's rename it to `includes_first_page`, which is
what that arg signifies - it tells us if the PDF we want to create
will contain the first page of a letter or not.

Corresponding changes are made in template-preview repo.