Add type property to Placeholder and refactor checks for conditional placeholders to use this property
To prepare for extending the placeholder syntax to include unsafe placeholders, we are introducing a type property to manage and be the single source of truth about what type of placeholder is being used to make extensions to the syntax easier to maintain and write.

We are assuming that a placeholder can only ever be of one type, and that union types are forbidden (e.g. unsafe and conditional placeholders are not allowed)