Make inheritence of annotations on SerialisedModel work on both the class and its instances
We do a neat thing where subclasses of `SerialisedModel` can inherit
annotations from each other, avoiding the need to redefine them all each
time.

This is a bit different to how dataclasses work, but it’s the same way
that methods and properties work.

However there’s currently a gotcha, in that it only works once the
object has been instantiated. So
`Custom.__annotations__ != Custom().__annotations`.

This leads to clunky workarounds like instantiating a throwaway object:
https://github.com/alphagov/notifications-admin/blob/c1d67ab2b991c3d7d452dedf8cd4b8fd8163f548/app/models/branding.py#L31

This commit moves the code which does the inheritence to a metaclass.
This means it runs at the time the class is defined, not when it’s
initialised.