Replace VML logo with image and text
The approach of using VML overlapping real
'GOV.UK' text has a problem if you're navigating
using tabbing, or for a moment when you click on
the logo. When this happens, Outlook inverts the
VML but leaves a gap in this where the text sits.
I can't find a way to prevent this and it looks
broken.

So, after designing a version of the logo in MS
Word and using some of the CSS applied in the HTML
when you export to HTML, I found we can have a
variant of the version we use for other email
clients, where the dot is in a font with a
circular period. This needs to be different enough
in Outlook, because it doesn't support
display: inline-block and needs the custom
mso-text-raise CSS property to position the dot in
the middle of the line, and padding on the parent
cells to align with the crown image.

I also made the following changes in this commit:
- the govuk_bg_colour variable was removed due to
  only being used once now
- stop setting background colour on central cells
  in banner as unnecessary and losing them helps
  print styles remove the banner colour
- tweak to line-height of dot, to get better
  overall position among popular email clients
- aria-label=GOV.UK added to span containing
  'GOV.UK' text to help Voiceover determine an
  accessible name for that part of the banner