RedisClient: add set_if_timestamp_newer method and script
this script allows an atomic compare-and-swap between two
msgpack-encoded payloads that contain float `timestamp`
fields.

why msgpack? because it has zero overhead when encoding
arbitrary strings, unlike e.g. json which may need to add
a lot of escaping characters. though this method & script
is agnostic to any contents of the msgpack value other than
the `timestamp` field, it is assumed that callers to this
method will want to further encode the payload of these
stored values so that the in-redis lua script (the
"critical section" in this scheme) doesn't have to parse
large & complex structures just to find the timestamp.

why not make the script responsible for the whole timestamp-
wrapping/encoding and decoding? to move as much of the
encoding and decoding work out of the critical section as
possible - the critical section effectively has exclusive
access to this key (and probably the whole redis instance)
for its whole runtime, so we really want to minimize the
work it has to do. with this scheme, the only extra work
the critical section needs to do is decoding two (hopefully
simple) msgpack values - and in cases where there is no
existing value, even that work will be omitted.

it's much more affordable performance-wise to have the
encoding/decoding work done by the clients because they
are distributed & horizontally scalable.

this tries to follow the interface of the RedisClient.set
method as much as possible, but omits the fancy options
like px, nx, xx, as it's not clear whether any of these
would make any sense in this context.