RequestCache: use tombstone mechanism instead of simply deleting keys
even with a delete-before, delete-after pattern, use of the
`delete` decorator contains a potential race condition. this is
because simply performing a delete when a cache is supposedly
invalidated doesn't take into account the fact that there
might already be another request in-flight that contains
now-stale data. when this other request "lands" it will
happily re-populate the cache with the stale (or at least
partially stale) data.

instead of performing a simple delete, overwrite the cache
value with a "tombstone" marker that contains a timestamp
of when the cache was supposedly invalidated (a `force_delete`
option remains for simply doing pure deletions, mostly to
allow for easy migration from previous utils versions, where
we will still want new app instances to be invalidating
the cache for old key schemes)

the `set` decorator is modified to treat these tombstones as
an empty value when encountered, and internally wraps its
stored values in a msgpack-encoded dict containing a
"pessimistic" timestamp for the generation of that value.
when setting a cache value it uses the new
`RedisClient.set_if_timestamp_newer`, which will skip storing
the new value if redis is found to contain another entry
with a newer timestamp than this "pessimistic" timestamp,
because this means that the new value may contain data from
before the deletion/invalidation was issued.

this also switches to using msgpack serialization instead of
json serialization for efficiency reasons detailed in the
addition of `RedisClient.set_if_timestamp_newer`.