RequestCache.delete_by_pattern: use tombstone mechanism instead of deletion
similar to the `delete` method, don't delete matching keys but
overwrite them with a timestamped tombstone marker.

sadly there isn't an easy way to bulk-set TTL for a number of
keys, so these tombstones will live as long as the values they
were replacing.

a `force_delete` option exists to cause a straightforward
deletion - this is intended to assist in rolling out this
new utils version, where we will be wanting both old and
new versions of the app to be able to invalidate both old
and new keys during deployment.