Update the letter delivery times
Starting soon, only first class letters will be delivered on a Saturday,
so this change reflects that when estimating the earliest and latest
delivery days.

The estimated times for second and economy mail are also changing.
We will show second class mail as taking 3-4 working days to deliver,
and economy mail as taking up to 6.