added `redis_client_with_live_instance` fixture to test against a live redis instance vs mock
RedisClient tests are currently run against a mocked version of the RedisClient. This is limited because it means we can test that the client and its methods are called with the right arguments, but we have no good way of testing the effect of operations on an actual redis db. Specifically, this change means we can now test that lua scripts behave as expected, we had no mechanism for doing this before.

Introduces a locking mechanism for the client to safeguard against race conditions. flask_redis doesn't appear to have an easy to use mechanism for using redis namespaces. To accomodate this, the redis client will run `FLUSHALL` every time it is called to ensure every test runs against a blank slate. We use xdist to run tests in parallel which introduces the rare chance of race conditions where a second test calls `FLUSHALL` on the db while the first test is midway through writing to redis. To guard against this you should add the decorator `@pytest.mark.xdist_group(name="group1")` to every test which uses this client to ensure that they all run on the same worker. pytest will raise an exception if someone fails to add this to their test and the tests run into this race condition.