Remove HTML and CSS fixing the main column width
As support moves to mso > 2019, we can use the
max-width and width 100% approach to make our
tables flex at different viewport sizes:

https://www.caniemail.com/features/css-max-width/

This also removes CSS in the `<head>` of the email
HTML targeting the main content column and fixing
its width.

These changes have been tested on emailonacid and
with emails to both 'new outlook' and 'outlook
classic'.