TMP UNTESTED RequestCache: add "internal" schema-versioning mechanism
this is to allow the expected format of the wrapped-method's
return value to evolve safely. when this format changes, the
schema_version kwarg's value should be incremented to avoid
mismatch problems that may occur during or shortly after
deployment of the new version.

this versioning scheme is "internal" in that the version is
contained *in* the cached value as opposed to a scheme
where the version might be included in the cache key instead.
this has the advantage that it forces two different app
versions to compete for the same cache key, rather than old
and new app versions having a different "view" of whether
a value is cached or not and the age of that cached value.
this becomes particularly important when considering how
to handle updated values and complete invalidation of stale
cached values for all possible app versions that could be
active at any time.

when a cached value with a mismatched schema version is
encountered, it is simply ignored and overwritten with the
inner function's return value. though it might seem that
the better thing to do would be to only defer to the higher
schema version (therefore always be "upgrading"), during
a slow or stuck deployment that could result in all of the
old versions of the app being prevented from using a cached
value at all. in a case where we had more old instances
than new instances this could cause a significant
performance problem. though an app always overwriting a
value with its own schema version could result in a
back-and-forth of continual overwriting between two app
versions, the probability of the cache holding a value
compatible with any particular app version is approximately
proportional to the proportional population of app
instances running it.