Update Zendesk client to allow topic fields to be populated
Our support form in Zendesk has different form fields for topics. We
haven't been populating these from the API, but we would like to start
doing so in a few cases. There are 22 topic fields available, each with
many potential values, so instead of putting every ID and value in the
Zendesk client, this change just allows the code to take a list of IDs
and values for the topic form fields and will pass those through as part
of the API call. Any wrong IDs or values will not stop the ticket from
being created, it simply won't populate a topic.