TMP UNTESTED RequestCache.set: make schema_version=0 behave like old-style RequestCache
to allow mostly-seamless rollout, an existing cache value that
can't be decoded by msgpack is assumed to be a json encoding
of an old-style RequestCache value and treated as having a
schema_version of 0.

similarly, if our local schema_version is specified as 0 (through
the schema_version arg), the payload will be stored back in redis
with a simple un-wrapped json-encoding. as such, when the local
schema_version is set to 0 it should behave identically to
old-style RequestCache and it should be possible to introduce this
wrapping-aware RequestCache seamlesslessly as schema_version 0,
before re-deploying it as schema_version=1 to enable the new
tombstone and schema-versioning functionality. interruption should
be similar to that caused by a regular schema_version bump being
rolled out.

the RequestCache.delete and RequestCache.delete_by_pattern methods
already have force_delete arguments that can be used to serve a
similar purpose.