Handle CSV files with thousands of empty columns
If you accidentally repeat your column headers across the whole sheet in Excel you will export a csv which looks like this:
```csv
phone number,name,name,name,name,name,
12345,,,,,,
67890,,,,,,
```
…except with `name` repeated 16,383 times.

So for a file with 100 rows we have to process 1,638,400 cells of data (and then trying to render that as an HTML table, if the code even makes it that far). The test I’ve added to this PR runs for at least a few minutes without halting on my machine.

This PR effectively cuts the file down to this:
```
phone number,name
12345,
67890,
```
…without breaking any of our existing expectations around how empty and duplicate columns should work.

This makes it processable.wip