Add a test-with-docker make task
This is a proposal to use the docker image we
already generate in `make-bootstrap-with-docker`
as another way to run the tests.

I don't think this is necessarily a better
approach than that described in the README but it
does provide some consistency with how concourse
does it and means the Dockerfile ends up as the
one description of how to run the tests.