Only read column headers from file once
Lots of bit of this code refer to the headers of the CSV file. Every
time they do this they put the whole contents into a `StringIO` object
and parse the first row of the file.

This commits uses Werkzeug’s `cached_property` utility to make it so we
only ever need to do this once per file.