Cache `RecipientCSV.duplicate_recipient_column_headers`
We call this once per row, per recipient column.

So if a file has a column called ‘phone number’ repeated many times, and
it has many rows, this can become a bit of a bottleneck.

Caching it means we only compute it once, no matter how many rows the
file has.